refactor(blocks): move edgeless type and remove useless type (#7113)
Move from surface block to edgeless root block.
