+++ packages/blocks/src/_common/components/embed-card/modal/embed-card-create-modal.ts
@@ -7,7 +7,7 @@ import { customElement, property, query, state } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 
 import type { EdgelessRootBlockComponent } from '../../../../root-block/edgeless/edgeless-root-block.js';
-import type { EdgelessElementType } from '../../../../surface-block/edgeless-types.js';
+import type { EdgelessElementType } from '../../../../root-block/edgeless/edgeless-types.js';
 import { Bound } from '../../../../surface-block/utils/bound.js';
 import { Vec } from '../../../../surface-block/utils/vec.js';
 import { EMBED_CARD_HEIGHT, EMBED_CARD_WIDTH } from '../../../consts.js';
