+++ packages/blocks/src/root-block/edgeless/edgeless-types.ts
@@ -1,0 +1,23 @@
+export type EdgelessBlockType =
+  | 'affine:frame'
+  | 'affine:note'
+  | 'affine:image'
+  | 'affine:attachment'
+  | 'affine:bookmark'
+  | 'affine:embed-github'
+  | 'affine:embed-youtube'
+  | 'affine:embed-figma'
+  | 'affine:embed-linked-doc'
+  | 'affine:embed-synced-doc'
+  | 'affine:embed-html'
+  | 'affine:embed-loom';
+
+export type EdgelessElementType =
+  | EdgelessBlockType
+  | 'shape'
+  | 'brush'
+  | 'connector'
+  | 'text'
+  | 'group'
+  | 'debug'
+  | 'mindmap';
