+++ packages/blocks/src/root-block/widgets/drag-handle/utils.ts
@@ -22,9 +22,9 @@ import {
   Rect,
 } from '../../../_common/utils/index.js';
 import type { ParagraphBlockModel } from '../../../paragraph-block/index.js';
-import type { EdgelessBlockType } from '../../../surface-block/index.js';
 import { Bound } from '../../../surface-block/index.js';
 import type { EdgelessRootBlockComponent } from '../../edgeless/edgeless-root-block.js';
+import type { EdgelessBlockType } from '../../edgeless/edgeless-types.js';
 import { isEmbedSyncedDocBlock } from '../../edgeless/utils/query.js';
 import {
   BLOCK_CHILDREN_CONTAINER_PADDING_LEFT,
