+++ packages/blocks/src/surface-block/index.ts
@@ -19,7 +19,6 @@ export {
   GRID_GAP_MAX,
   GRID_GAP_MIN,
 } from './consts.js';
-export { type EdgelessBlockType } from './edgeless-types.js';
 export { ElementModel } from './element-model/base.js';
 export { BrushElementModel } from './element-model/brush.js';
 export {
